+++ business_gen.py
@@ -1,16 +1,15 @@
 import requests
 from dotenv import load_dotenv
 import json
-from bs4 import BeautifulSoup
 import os
 import re
 import urllib.parse
 import boto3
 import csv
-from bs4 import BeautifulSoup
 
-from maps import geocode
-from model import business
+from maps.geocode import call_google
+from model.business import create_businesses
+from scraper.bs import scrape
 
 
 load_dotenv()
@@ -23,14 +22,6 @@ QUEUE_URL = os.getenv("QUEUE_URL")
 # NUM_DIVISIONS = 5  # Number of subdivisions in each dimension (change as needed)
 requests.packages.urllib3.disable_warnings()
 
-
-# class BusinessData(BaseModel):
-#     business_name: str
-#     url: str = ""  # Default empty string
-#     email: str = ""
-#     web_content: str = ""
-
-# TODO: add type hints
 
 
 def save_to_initial_queue(data, queue_url):
@@ -46,79 +37,79 @@ def save_to_initial_queue(data, queue_url):
     send_to_sqs(initial_data, queue_url)
 
 
-def extract_website_content(url):
-    headers = {
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
-    }
+# def extract_website_content(url):
+#     headers = {
+#         "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
+#     }
 
-    if not url or not url.startswith("http"):
-        return ""
+#     if not url or not url.startswith("http"):
+#         return ""
 
-    try:
-        response = requests.get(url, headers=headers, verify=False, timeout=10)
-        response.raise_for_status()
+#     try:
+#         response = requests.get(url, headers=headers, verify=False, timeout=10)
+#         response.raise_for_status()
 
-        soup = BeautifulSoup(response.content, "html.parser")
+#         soup = BeautifulSoup(response.content, "html.parser")
 
-        # Remove script and style elements
-        for script in soup(["script", "style"]):
-            script.extract()
+#         # Remove script and style elements
+#         for script in soup(["script", "style"]):
+#             script.extract()
 
-        # Get text content
-        text = soup.get_text()
+#         # Get text content
+#         text = soup.get_text()
 
-        # Break into lines and remove leading and trailing whitespace
-        lines = (line.strip() for line in text.splitlines())
+#         # Break into lines and remove leading and trailing whitespace
+#         lines = (line.strip() for line in text.splitlines())
 
-        # Drop blank lines
-        clean_lines = list(line for line in lines if line)
+#         # Drop blank lines
+#         clean_lines = list(line for line in lines if line)
 
-        return "\n".join(clean_lines)
+#         return "\n".join(clean_lines)
 
-    except Exception as e:
-        print(f"Error extracting content from {url}: {e}")
-        return ""
+#     except Exception as e:
+#         print(f"Error extracting content from {url}: {e}")
+#         return ""
 
 
-def extract_email_from_website(url, depth=1):
-    if not url:
-        return None
-    headers = {
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
-    }
-    # If the URL doesn't start with 'http' (could be http or https), then prepend it with 'http://'
-    if not url.startswith("http"):
-        url = "http://" + url
-    try:
-        response = requests.get(url, headers=headers, verify=False)
-        response.raise_for_status()
-        soup = BeautifulSoup(response.text, "html.parser")
-        # First, try to find mailto links
-        mailtos = soup.select("a[href^=mailto]")
-        for i in mailtos:
-            return i["href"].replace("mailto:", "")
-        # If that doesn't find an email, try searching the text using regex
-        email_pattern = r"[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}"
-        match = re.search(email_pattern, soup.get_text())
-        if match:
-            return match.group(0)
-        # If depth allows and we haven't found an email, follow potential "contact" links
-        if depth > 0:
-            contact_links = soup.select(
-                'a[href*="contact"], a[href*="email"], a[href*="Contact"], a[href*="Email"]'
-            )
-            for link in contact_links:
-                href = link.get("href")
-                if href:
-                    # Build a full URL if it's a relative link
-                    if not href.startswith(("http://", "https://")):
-                        href = urllib.parse.urljoin(url, href)
-                    email = extract_email_from_website(href, depth - 1)
-                    if email:
-                        return email
-    except Exception as e:  # This will catch all exceptions
-        print(f"Error extracting email from {url}: {e}")
-        return None
+# def extract_email_from_website(url, depth=1):
+#     if not url:
+#         return None
+#     headers = {
+#         "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
+#     }
+#     # If the URL doesn't start with 'http' (could be http or https), then prepend it with 'http://'
+#     if not url.startswith("http"):
+#         url = "http://" + url
+#     try:
+#         response = requests.get(url, headers=headers, verify=False)
+#         response.raise_for_status()
+#         soup = BeautifulSoup(response.text, "html.parser")
+#         # First, try to find mailto links
+#         mailtos = soup.select("a[href^=mailto]")
+#         for i in mailtos:
+#             return i["href"].replace("mailto:", "")
+#         # If that doesn't find an email, try searching the text using regex
+#         email_pattern = r"[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}"
+#         match = re.search(email_pattern, soup.get_text())
+#         if match:
+#             return match.group(0)
+#         # If depth allows and we haven't found an email, follow potential "contact" links
+#         if depth > 0:
+#             contact_links = soup.select(
+#                 'a[href*="contact"], a[href*="email"], a[href*="Contact"], a[href*="Email"]'
+#             )
+#             for link in contact_links:
+#                 href = link.get("href")
+#                 if href:
+#                     # Build a full URL if it's a relative link
+#                     if not href.startswith(("http://", "https://")):
+#                         href = urllib.parse.urljoin(url, href)
+#                     email = extract_email_from_website(href, depth - 1)
+#                     if email:
+#                         return email
+#     except Exception as e:  # This will catch all exceptions
+#         print(f"Error extracting email from {url}: {e}")
+#         return None
 
 
 # adding csv to prove we are obtaining them
@@ -193,8 +184,10 @@ def main():
     # min_lat, max_lat = lat - delta, lat + delta
     # min_lng, max_lng = lng - delta, lng + delta
     # data = geocode.get_places(query, min_lat, max_lat, min_lng, max_lng)
-    data = geocode.call_google()
-    businesses = business.create_businesses(data)
+    data = call_google()
+    businesses = create_businesses(data)
+    print(f"scraping only {businesses[0].name}")
+    lines = scrape(businesses[0])
     # save_to_initial_queue(data, QUEUE_URL)
     # save_to_sqs(data, QUEUE_URL)
     # # added csv to prove we are getting emails
