+++ model/business.py
@@ -1,10 +1,13 @@
 from typing import List
 from pydantic import BaseModel
 
+from scraper.bs import extract_website_content
+from scraper.bs import extract_email_from_website
+
 
 class Business(BaseModel):
     name: str
-    website: str = ""  # Default empty string
+    website: str = ""
     email: str = ""
     web_content: str = ""
 
@@ -13,10 +16,13 @@ def create_businesses(data) -> List[Business]:
     businesses = []
 
     for item in data:
+        # adding name and website to Business
         business = Business(**item)
         print("adding business")
-        print(business.name)
-        print(business.website)
+        # adding web_content to Business
+        business.web_content = extract_website_content(business.website)
+        # adding email to Business
+        business.email = extract_email_from_website(business.website)
         businesses.append(business)
     
     return businesses
